services/video-processing/requirements.txt
# Video Processing Dependencies
# Install with: pip install -r requirements.txt

# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Video Processing
ffmpeg-python==0.2.0

# Speech-to-Text (Whisper for multilingual support including Mongolian)
openai-whisper==20231117
# Alternative: faster-whisper for better performance
# faster-whisper==0.10.0

# Supabase integration
supabase==2.3.4

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
aiofiles==23.2.1
httpx==0.26.0
